[refactor♻️] 비즈니스 레이어 추상화(#14)
